+++ index.js
@@ -6,6 +6,38 @@ const { setSvgBbox, validSVG } = require("./src/svg.js");
 const PLUGIN_NAME = "remark-mermaid-dataurl";
 
 /**
+ * Adds custom `remark-mermaid-dataurl` defaults to a mermaid config file.
+ *
+ * Sets `useMaxWidth` to `false` by default for better Markdown SVGs.
+ *
+ * @param {object} mermaidConfig - Warning, this is modified by this function.
+ * @returns {object} The input object with some default vars modified.
+ */
+function addDefaultConfig(mermaidConfig) {
+  const GRAPHS_TO_DISABLE_MAX_WIDTH = [
+    "flowchart",
+    "sequence",
+    "gantt",
+    "journey",
+    "class",
+    "state",
+    "er",
+    "pie",
+    "requirement",
+    "c4",
+  ];
+  for (const graphType of GRAPHS_TO_DISABLE_MAX_WIDTH) {
+    mermaidConfig[graphType] = {
+      // if this is true (default), SVG will use up the entire width of the markdown
+      // document, which is much much too wide
+      useMaxWidth: false,
+      ...mermaidConfig[graphType],
+    };
+  }
+  return mermaidConfig;
+}
+
+/**
  * Converts CLI args to {@link parseMMD} options.
  *
  * Required for backwards compatibility.
@@ -37,7 +69,7 @@ async function convertMermaidKwargsToParseMMDOpts({
   }
 
   return {
-    mermaidConfig,
+    mermaidConfig: addDefaultConfig(mermaidConfig),
     backgroundColor,
     myCSS,
     pdfFit,